Move unused result helpers into partial
